# fix(server): prohibit the length headers on successful CONNECT