@seanmonstar 2019-03-18
Thanks! Out of curiosity, do you know if the additional branches affect performance much? Likely the ones with least noise are in `proto::h1::role` at the bottom of the file.