@ubnt-intrepid 2019-03-19
The benchmark result on my environment:

master:

```
test proto::h1::role::tests::bench_parse_incoming               ... bench:       3,199 ns/iter (+/- 1,142) = 274 MB/s
test proto::h1::role::tests::bench_parse_short                  ... bench:         282 ns/iter (+/- 10) = 141 MB/s
test proto::h1::role::tests::bench_server_encode_headers_preset ... bench:         246 ns/iter (+/- 13) = 439 MB/s
test proto::h1::role::tests::bench_server_encode_no_headers     ... bench:          34 ns/iter (+/- 0) = 2235 MB/s
```

this PR:

```
test proto::h1::role::tests::bench_parse_incoming               ... bench:       3,229 ns/iter (+/- 1,014) = 271 MB/s
test proto::h1::role::tests::bench_parse_short                  ... bench:         286 ns/iter (+/- 6) = 139 MB/s
test proto::h1::role::tests::bench_server_encode_headers_preset ... bench:         256 ns/iter (+/- 20) = 421 MB/s
test proto::h1::role::tests::bench_server_encode_no_headers     ... bench:          34 ns/iter (+/- 0) = 2235 MB/s
```