@seanmonstar 2019-03-19
Oh yea, so it does affect slightly `bench_server_encode_headers_preset`. Since this condition should be very rare, I think we can structure the change so that the normal case doesn't see any extra branches.