@ubnt-intrepid 2019-03-20
Fixed the patch to make the number of branch to be the same as the original.

<details>
<summary>benchmark result</summary>
master:

```
test proto::h1::role::tests::bench_parse_incoming               ... bench:       3,176 ns/iter (+/- 74) = 276 MB/s
test proto::h1::role::tests::bench_parse_short                  ... bench:         280 ns/iter (+/- 4) = 142 MB/s
test proto::h1::role::tests::bench_server_encode_headers_preset ... bench:         259 ns/iter (+/- 30) = 416 MB/s
test proto::h1::role::tests::bench_server_encode_no_headers     ... bench:          34 ns/iter (+/- 0) = 2235 MB/s
```

this PR:

```
test proto::h1::role::tests::bench_parse_incoming               ... bench:       3,193 ns/iter (+/- 107) = 274 MB/s
test proto::h1::role::tests::bench_parse_short                  ... bench:         280 ns/iter (+/- 8) = 142 MB/s
test proto::h1::role::tests::bench_server_encode_headers_preset ... bench:         245 ns/iter (+/- 20) = 440 MB/s
test proto::h1::role::tests::bench_server_encode_no_headers     ... bench:          34 ns/iter (+/- 0) = 2235 MB/s
```
</details>